Add files via upload
The "distjs" is a JavaScript file that has to be pasted in the "simulation script editor" window found in the "tools" tab. Then the code has to be made "Activate" from the "Run" tab of script editor window and the output of it has to be copied into the "coor" text file. The provided coor.txt is a sample file for example purpose. Then the euclidDist file has to be executed and server has to be chosen to get distance from the nodes,